added 2 controllers

+++ Application/Helpers/FilterParameters.cs
@@ -11,13 +11,13 @@ public record FilterParameters
     int pageSize,
     int pageNumber,
     string? title,
-    decimal? minPrice,
-    decimal? maxPrice,
+    double? minPrice,
+    double? maxPrice,
     bool orderByTitle = true)
 {
     public string Title = title ?? string.Empty;
-    public decimal MinPrice = minPrice ?? 0;
-    public decimal MaxPrice = maxPrice ?? decimal.MaxValue;
+    public double MinPrice = minPrice ?? 0;
+    public double MaxPrice = maxPrice ?? double.MaxValue;
     public bool OrderByTitle = orderByTitle;
     public int PageNumber = pageNumber;
     public int PageSize = pageSize;
